requirements.txt
Adafruit-ADS1x15==1.0.2
Adafruit-GPIO==1.0.3
Adafruit-PureIO==1.1.11
appdirs==1.4.4
astroid==2.4.2
distlib==0.3.7
filelock==3.12.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.26.0
opencv-python==4.6.0.66
pexpect==4.8.0
Pillow==9.5.0
platformdirs==3.10.0
psutil==5.7.2
ptyprocess==0.7.0
pylint==2.5.3
pyLoRa==0.3.1
RPi.GPIO==0.7.1
six==1.15.0
spidev==3.6
toml==0.10.1
virtualenv==20.24.5
wrapt==1.12.1
